add in memory gateway
